Add LazyResizableImage to the Design system module
